rename autoloader file
